+++ src/game/Map.h
@@ -204,7 +204,6 @@ class MANGOS_DLL_SPEC Map : public GridRefManager<NGridType>
         {
             return i_mapEntry && i_mapEntry->Instanceable();
         }
-        // NOTE: this duplicate of Instanceable(), but Instanceable() can be changed when BG also will be instanceable
         bool IsDungeon() const
         {
             return i_mapEntry && i_mapEntry->IsDungeon();
@@ -216,6 +215,10 @@ class MANGOS_DLL_SPEC Map : public GridRefManager<NGridType>
         bool IsBattleGround() const
         {
             return i_mapEntry && i_mapEntry->IsBattleGround();
+        }
+        bool IsContinent() const
+        {
+            return i_mapEntry && i_mapEntry->IsContinent();
         }
 
         // can't be NULL for loaded map